requirements.txt
faster_whisper==1.1.0
ffmpeg_python==0.2.0
openai==1.86.0
OpenCC==1.1.1
python_docx==1.1.2
PyYAML==6.0.2
PyYAML==6.0.2
rich==14.0.0
thefuzz==0.22.1
volcengine_python_sdk==1.0.115
